+++ backend/packages/workspaces/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/workspaces
 
+## 0.27.14
+
+### Patch Changes
+
+- Updated dependencies [9b3473a]
+  - @cat-factory/contracts@0.294.0
+  - @cat-factory/kernel@0.286.2
+
 ## 0.27.13
 
 ### Patch Changes